docs: add how-to guide on scaling with LO
This notebook contains a basic example of how gate cutting can be useful when dealing with periodic problems. This is the basic version that doesn't use any error mitigation method, and simply deals with cutting the periodic gate. This notebook does not use the Qiskit Primitives V2 because there is still no update on Circuit knitting toolbox to use them.