docs: add how-to guide on wire-cutting
This is a how-to guide on wire cutting that was presented at the APAC Practitioner's Forum. There are two known drawbacks to this - (i) for simulation it uses cloud simulators since there are some issues with Qiskit which prevents us to use local simulator with circuit-knitting-toolbox at this moment; and (ii) circuit-knitting-toolbox which supports Primitives V2 is not yet publicly available. So we use Primitives V1 only.